* GAM-product-seeds: (43 commits)
  Gam frontend sentry
  Sentry Error Handeling
  User Soft Deletion
  Clean up of the rule
  Fix of the deletes
  Fix of the specifications post to macht for that product_id
  Fix of the search with no query
  Fixes to the mix
  Replaced all 'asset' thingies with 'mix' for mix.webpack versioning
  Remove of the mix-manifest
  improved factories with states
  removed var dump
  added image generator to productfactory
  Basic Products seed
  Add of the product nont debuged
  Fix of the bug if there is not three cars in database
  Fix of the styling
  Committing db:seed changes
  Add of the coming_soon status
  almost done with emails
  ...

+++ app/Http/Controllers/Admin/Products/ProductsBrandsController.php
@@ -6,7 +6,8 @@ use App\Facades\ImageUtil;
 use App\Http\Controllers\Controller;
 use App\Http\Requests\CreateBrandRequest;
 use App\Http\Requests\UpdateBrandRequest;
-use App\Model\ProductsBrands;
+use App\Models\ProductsBrands;
+use App\Models\Products;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Storage;
 
@@ -60,15 +61,32 @@ class ProductsBrandsController extends Controller
         return redirect()->route('admin.products.brands.index');
     }
 
+    private function guardDelete(ProductsBrands $brand)
+    {
+        $models_count = $brand->models()->count();
+        $products_with_brand = Products::where('brand_id', $brand->id)->count();
+        if ($products_with_brand !== 0) {
+            flash_message('Could not delete this brand, because it is currently in use in one or more products and has models.', 'error');
+            return true;
+        } elseif ($models_count !== 0) {
+            flash_message('Could not delete this brand, because it is still has models.', 'error');
+            return true;
+        }
+        return false;
+    }
+
     public function delete(ProductsBrands $brand)
     {
+        if($this->guardDelete($brand)){
+            return redirect()->route('admin.products.brands.index');
+        }
+
         try {
             $brand->delete();
             flash_message('The brand deleted successfully.', 'warning');
         } catch (\Exception $e) {
-            if ($e->errorInfo[1] == 1451) {
-                flash_message('Could not delete this brand, because it is currently in use in one or more products.', 'error');
-            }
+            flash_message('Could not delete this brand for unknown error', 'error');
+            // TODO implement an exception handel
         }
 
         return redirect()->route('admin.products.brands.index');
@@ -77,5 +95,14 @@ class ProductsBrandsController extends Controller
     public function getBrandsName()
     {
         return ProductsBrands::orderBy('title', 'ASC')->get();
+    }
+
+    public function getBrandNamesArray() {
+        $brands = $this->getBrandsName();
+        $number_of_brands = count($brands);
+        for ( $i = 0; $i < $number_of_brands; $i++) {
+            $brands_list[] = $brands[$i]['title'];
+        }
+        return $brands_list;
     }
 }
